% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks.R
\name{add_external_dependency}
\alias{add_external_dependency}
\title{Add external file to use as dependency}
\usage{
add_external_dependency(
  analysis,
  dependency_file,
  dependency_name = fs::path_file(dependency_file)
)
}
\arguments{
\item{analysis}{analysis object (as created with new_analysis)}

\item{dependency_file}{character string. Specifies the path of the external dependency}

\item{dependency_name}{character string. Name under which above path is available as parameter to notebooks.}
}
\description{
Add external file to use as dependency
}
