% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_notebooks.R
\name{nf_text}
\alias{nf_text}
\title{Return the full .nf content as a single character string}
\usage{
nf_text(analysis, analysis_name = analysis$name)
}
\arguments{
\item{analysis}{analysis object}

\item{analysis_name}{character. Used in the file header comment.}
}
\value{
length-n character vector.
}
\description{
Return the full .nf content as a single character string
}
