% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_notebooks.R
\name{nf_text_single}
\alias{nf_text_single}
\title{Return the full .nf content as a single character string}
\usage{
nf_text_single(
  analysis_name = analysis$name,
  process_blocks,
  call_lines,
  includes = character(0)
)
}
\arguments{
\item{analysis_name}{character. Used in the file header comment.}

\item{process_blocks}{character vector, each containing a process block.}

\item{call_lines}{character vector, each containing a workflow call line.}

\item{includes}{named character vector, each containing a output to include named by module.}
}
\value{
length-1 character vector.
}
\description{
Return the full .nf content as a single character string
}
