% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_notebooks.R
\name{write_makefile}
\alias{write_makefile}
\title{Write makefile}
\usage{
write_makefile(
  analysis,
  analysis_name = analysis$name,
  makefile = paste0(analysis_name, ".mk"),
  rmarkdown_params = NULL
)
}
\arguments{
\item{analysis}{analysis object}

\item{analysis_name}{character string. Name of make targets}

\item{makefile}{file path. Name of the makefile}

\item{rmarkdown_params}{list of additional parameters for rmarkdown::render}
}
\description{
Write makefile
}
