% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_notebooks.R
\name{add_notebook}
\alias{add_notebook}
\title{Add notebook to analysis}
\usage{
add_notebook(
  analysis,
  notebook_file,
  products = character(0),
  dependencies = list(),
  params = list(),
  notebook_name = fs::path_sanitize(fs::path_ext_remove(notebook_file))
)
}
\arguments{
\item{analysis}{analysis object}

\item{notebook_file}{file path of the notebook to add to analysis (relative to analysis$notebook_dir)}

\item{products}{(named) character vector specifying file paths of side products of the notebook, they are available as parameters to subsequent notebooks by their name}

\item{dependencies}{either a named list of character vectors or a named character vector specifying file paths of files required by the notebook. Dependencies matching parameters will be used as such}

\item{params}{named list of parameters to supply to the notebook}

\item{notebook_name}{character string to use as name of the notebook under which its out_dir will available as parameter to subsequent notebooks}
}
\description{
Add notebook to analysis
}
