% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysismaker.R
\docType{package}
\name{analysismaker}
\alias{analysismaker}
\title{Helper package to create make files for .RMD files with dependencies}
\description{
This package halps to create makefiles to render interdependent notebooks with parameters.
}
