% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_notebooks.R
\name{new_analysis}
\alias{new_analysis}
\title{Create new analysis object}
\usage{
new_analysis(
  name = "analysis",
  notebook_dir = "notebooks",
  out_dir = "results",
  out_dir_human = fs::path("results_human", name)
)
}
\arguments{
\item{name}{character string. Specifies name of the analysis (used in output file names)}

\item{notebook_dir}{base directory of the notebook files}

\item{out_dir}{base directory for the result files}

\item{out_dir_human}{base directory for links to the result files with human readable names}
}
\description{
Create new analysis object
}
